+++ app/src/main/java/woowacourse/movie/feature/setting/SettingFragment.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.feature
+package woowacourse.movie.feature.setting
 
 import android.os.Bundle
 import android.view.LayoutInflater
